Add files via upload
These ones are html files because i am using wordpress to develop them faster, but anyway anyone of you guys can modify the html and php codes as you guys please since they are simple and they cut to the cheese in matters of website designs, and i beg you guys that lend me a hand on  the best choice of hosting and domain when working with wordpress as i am beginning my job on this. a shout out to everyone encouraged to bring something to the table about these topics and projects.